Bug fixes & updates to improve code re-use.
Added utility methods and class.

+++ src/main/java/com/hsjawanda/gaeobjectify/data/GaeDataUtil.java
@@ -207,6 +207,8 @@ public class GaeDataUtil {
 	}
 
 	public static <T> void deferredDeleteEntity(T t) {
+		if (null == t)
+			return;
 		ofy().defer().delete().entity(t);
 	}
